requirements.txt
python
streamlit
opencv-python
pillow
numpy